requirements.txt
numpy==2.3.1
pillow==11.3.0
pydub==0.25.1
scipy==1.16.0
tqdm==4.67.1
